requirements.txt
langchain-openai>=0.0.4
langchain>=0.3.25
langgraph>=0.4.3
langgraph-cli>=0.2.12
langchain-core>=0.3.64
langgraph-api>=0.2.45
python-dotenv>=1.1.0